(refractor) Code refractor
