--- server/services/reports/src/main/java/com/recykling/report/constans/EmployeeConstants.java
@@ -1,13 +1,0 @@
-package com.recykling.report.constans;
-/**
- * @author WiniaR21
- */
-public class EmployeeConstants {
-    private EmployeeConstants(){
-        //  Restrict instantiation
-    }
-    public static final String  STATUS_200 = "200";
-    public static final String  MESSAGE_200 = "Request processed successfully";
-    public static final String  STATUS_201 = "201";
-    public static final String  MESSAGE_201 = "Employee created successfully";
-}
