--- server/services/reports/src/main/java/com/recykling/report/dto/ResponseDto.java
@@ -1,13 +1,0 @@
-package com.recykling.report.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-/**
- * @author WiniaR21
- */
-@Data @AllArgsConstructor
-public class ResponseDto {
-    private String statusCode;
-    private String statusMsg;
-}
