--- server/services/reports/src/main/java/com/recykling/report/entity/reports/ReportBase.java
@@ -1,20 +1,0 @@
-package com.recykling.report.entity.reports;
-
-import com.recykling.report.valueObjects.EmployeesCount;
-import com.recykling.report.valueObjects.ReportDate;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-/**
- * @author WiniaR21
- */
-@MappedSuperclass
-@Getter
-@Setter
-@ToString
-public class ReportBase {
-    private ReportDate reportDate;
-    private EmployeesCount employeesCount;
-}
