+++ server/services/reports/src/main/java/com/recykling/report/exception/dto/ErrorResponseDto.java
@@ -1,0 +1,18 @@
+package com.recykling.report.exception.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import org.springframework.http.HttpStatus;
+
+import java.time.LocalDateTime;
+
+/**
+ * @author WiniaR21
+ */
+@Data @AllArgsConstructor
+public class ErrorResponseDto {
+    private String apiPath;
+    private HttpStatus errorCode;
+    private String errorMessage;
+    private LocalDateTime errorTime;
+}
