+++ server/services/reports/src/main/java/com/recykling/report/exception/dto/ResponseDto.java
@@ -1,0 +1,13 @@
+package com.recykling.report.exception.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+/**
+ * @author WiniaR21
+ */
+@Data @AllArgsConstructor
+public class ResponseDto {
+    private String statusCode;
+    private String statusMsg;
+}
