--- server/services/reports/src/main/java/com/recykling/report/repository/AggregatesWithoutOilWeightsRepository.java
@@ -1,9 +1,0 @@
-package com.recykling.report.repository;
-
-import com.recykling.report.entity.reports.AggregatesWithoutOilWeights;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AggregatesWithoutOilWeightsRepository extends JpaRepository<AggregatesWithoutOilWeights, Long> {
-}
