--- server/services/reports/src/main/java/com/recykling/report/repository/UrtReportHistoryRepository.java
@@ -1,10 +1,0 @@
-package com.recykling.report.repository;
-
-import com.recykling.report.entity.UrtReportHistory;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UrtReportHistoryRepository extends JpaRepository<UrtReportHistory, Long> {
-
-}
