+++ server/services/reports/src/main/java/com/recykling/report/urtReport/urtAggregatesWithoutOulWeights/AggregatesWithoutOilWeights.java
@@ -1,0 +1,27 @@
+package com.recykling.report.urtReport.urtAggregatesWithoutOulWeights;
+
+import com.recykling.report.urtReport.UrtReport;
+import com.recykling.report.valueObjects.Weight;
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "urt_aggregates_without_oil_weights")
+@NoArgsConstructor
+@Getter
+public class AggregatesWithoutOilWeights {
+    @Id
+    @GeneratedValue(strategy = GenerationType.SEQUENCE)
+    @Column(name = "aggregates_without_oil_weights_id")
+    private Long aggregatesWithoutOilWeightsId;
+    private Integer weight;
+    @ManyToOne
+    @JoinColumn(name="urt_report_id", nullable = false)
+    private UrtReport urtReport;
+
+    public AggregatesWithoutOilWeights(Weight weight, UrtReport urtReport) {
+        this.weight = weight.weight();
+        this.urtReport = urtReport;
+    }
+}
